Create folder ArrayFieldTemplate and move all related components to it. Create ItemForm component and remove isOpen logic from ArrayFieldTemplateItem.
