+++ react/components/form/ArrayFieldTemplate/AddButton.tsx
@@ -2,20 +2,17 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 import { ButtonWithIcon } from 'vtex.styleguide'
 
-import AddIcon from '../icons/AddIcon'
+import AddIcon from '../../icons/AddIcon'
 
 interface Props {
   onClick: (event: Event) => void
 }
 
+const icon = <AddIcon />
+
 const AddButton: React.FC<Props> = ({ onClick }) => (
   <div className="ba b--muted-4 br2 mb3">
-    <ButtonWithIcon
-      block
-      icon={<AddIcon />}
-      onClick={onClick}
-      variation="tertiary"
-    >
+    <ButtonWithIcon block icon={icon} onClick={onClick} variation="tertiary">
       <FormattedMessage
         defaultMessage="Add"
         id="admin/pages.editor.components.arrayTemplate.button.add"
