+++ react/components/form/ArrayFieldTemplate/ArrayFieldTemplateItem/PreviewOverlay.tsx
